removed valuescontroller as it was not being used, removed moveregister.cs and moveuser.cs as they were moved to the movecontroller. merged valuescontroller, variancecontroller, getclosecount, and createtransactioncontroller together into transactioncontroller
