+++ CIS424_API/CIS424_API.csproj
@@ -186,6 +186,7 @@
     <Compile Include="Areas\HelpPage\SampleGeneration\TextSample.cs" />
     <Compile Include="Areas\HelpPage\XmlDocumentationProvider.cs" />
     <Compile Include="Controllers\AuthenticateUserController.cs" />
+    <Compile Include="Controllers\TransactionController.cs" />
     <Compile Include="Controllers\BaseApiController.cs" />
     <Compile Include="Controllers\CreateUserController.cs" />
     <Compile Include="Controllers\DisableController.cs" />
@@ -195,11 +196,7 @@
     <Compile Include="Controllers\RegisterController.cs" />
     <Compile Include="Controllers\StoreController.cs" />
     <Compile Include="Controllers\TransferController.cs" />
-    <Compile Include="Controllers\ValuesController.cs" />
-    <Compile Include="Controllers\CreateTransactionController.cs" />
-    <Compile Include="Controllers\VarianceController.cs" />
     <Compile Include="Controllers\ViewUsersController.cs" />
-    <Compile Include="Controllers\GetExpectedCount.cs" />
     <Compile Include="Controllers\EditUser.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
